+++ lua/qmk/parse/init.lua
@@ -1,8 +1,10 @@
 local E = require('qmk.errors')
-local parse_qmk = require('qmk.parse.qmk')
 local check = require('qmk.utils').check
 
-local M = {}
+local M = {
+	qmk = require('qmk.parse.qmk'),
+	zmk = require('qmk.parse.zmk'),
+}
 
 ---assert all keymaps don't overlap with the declaration itself
 ---@param keymaps qmk.Keymaps
@@ -26,10 +28,10 @@ end
 ---currenly only supports qmk keymaps, but in theory could support anything that parses to a qmk.Keymaps
 ---@param content string
 ---@param options qmk.Config
----@param parser? fun(content: string, options: qmk.Config): qmk.Keymaps
+---@param parser fun(content: string, options: qmk.Config): qmk.Keymaps
 ---@return qmk.Keymaps
 function M.parse(content, options, parser)
-	local board_parser = parser or parse_qmk
+	local board_parser = parser
 	local keymaps = board_parser(content, options)
 	validate(keymaps)
 	return keymaps
