+++ lua/qmk/parse/zmk/init.lua
@@ -39,6 +39,7 @@ local function get_keymaps(root, content)
 					local text = (ts.get_node_text(child, content))
 
 					if child:type() == 'reference' or vim.startswith(text, '_') then
+						-- this is the start of a new key code
 						vim.list_extend(keys, { { text } })
 					else
 						vim.list_extend(keys[#keys], { text })
